# openapi.yaml
openapi: "3.1.0"
info:
  title: "UPC API Institucional — Módulo Académico"
  description: |
    RESTful integration API for **Universidad Popular del Cesar (UPC)** institutional systems.

    ## Overview
    This API provides programmatic access to academic data from UPC's ACADEMUSOFT system,
    including programmes, subjects, periods, student enrollments, and final grades.

    ## Authentication
    All protected endpoints require a valid **OAuth 2.0 Bearer token** issued by the UPC
    Keycloak identity provider. Two flows are supported:
    - **Client Credentials** (machine-to-machine integrations)
    - **Authorization Code + PKCE** (user-delegated access)

    Tokens are validated against the Keycloak JWKS endpoint. The `scope` claim must include
    the required scope for each endpoint.

    ## Rate Limiting
    - Production: 1 000 requests / minute per client
    - Sandbox: 200 requests / minute per client

    HTTP `429 Too Many Requests` is returned when the limit is exceeded.

    ## Environments
    | Environment | Base URL |
    |---|---|
    | Production | `https://api.unicesar.edu.co/v1` |
    | Sandbox | `https://api-sandbox.unicesar.edu.co/v1` |

  version: "1.0.0"
  contact:
    name: "Dirección de Tecnologías de la Información — UPC"
    email: "api-soporte@unicesar.edu.co"
    url: "https://www.unicesar.edu.co"
  license:
    name: "Proprietary"
    url: "https://www.unicesar.edu.co/legal"
servers:
  - url: "https://api.unicesar.edu.co/v1"
...
components:
  securitySchemes:

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by the UPC Keycloak identity provider.
        Validate the token against:
        `https://auth.unicesar.edu.co/realms/upc/protocol/openid-connect/certs`

    OAuth2ClientCredentials:
      type: oauth2
      description: "Machine-to-machine (M2M) flow"
      flows:
        clientCredentials:
          tokenUrl: "https://auth.unicesar.edu.co/realms/upc/protocol/openid-connect/token"
          scopes:
            "academico:read": "Read academic programs, subjects and periods"
            "estudiantes:read": "Read student enrollment and grade data"
            "docentes:read": "Read faculty / teaching staff data"
            "administrativo:read": "Read administrative data"
            "sgdea:read": "Read SGDEA document management data"
            "soporte:read": "Read support / helpdesk data"
            "eventos:write": "Create and update institutional events"
            "admin:read": "Read administrative API metrics and health details"

    OAuth2AuthorizationCode:
      type: oauth2
      description: "User-delegated access via Authorization Code + PKCE"
      flows:
        authorizationCode:
          authorizationUrl: "https://auth.unicesar.edu.co/realms/upc/protocol/openid-connect/auth"
          tokenUrl: "https://auth.unicesar.edu.co/realms/upc/protocol/openid-connect/token"
          scopes:
            "academico:read": "Read academic programs, subjects and periods"
            "estudiantes:read": "Read student enrollment and grade data"
            "docentes:read": "Read faculty / teaching staff data"
            "administrativo:read": "Read administrative data"
            "sgdea:read": "Read SGDEA document management data"
            "soporte:read": "Read support / helpdesk data"
...
  schemas:
    ErrorResponse:
      type: object
      description: "Standard error response returned on all 4xx and 5xx responses"
      required: [error, message, timestamp, requestId]
      properties:
        error:
          type: string
          description: "Machine-readable error code"
          example: "UNAUTHORIZED"
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_ERROR
            - RATE_LIMIT_EXCEEDED
            - INTERNAL_ERROR
        message:
          type: string
          description: "Human-readable description of the error"
          example: "Token has expired"
        timestamp:
          type: string
          format: date-time
          description: "UTC timestamp of the error"
          example: "2025-07-10T14:30:00.000Z"
        requestId:
          type: string
          format: uuid
          description: "Unique identifier for this request (use when contacting support)"
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
...
    Inscripcion:
      type: object
      description: "A student's enrollment in a course for a given period"
      required:
        - codigoAsignatura
        - nombreAsignatura
        - codigoPrograma
        - periodo
        - creditos
        - grupo
        - docente
        - horario
        - salon
        - estadoInscripcion
      properties:
        codigoAsignatura:
          type: string
          example: "IS-301"
        nombreAsignatura:
          type: string
          example: "Ingeniería de Software I"
        codigoPrograma:
          type: string
          example: "ING-SIS-01"
        periodo:
          type: string
          example: "2025-2"
        creditos:
          type: integer
          example: 3
        grupo:
          type: string
          example: "A"
        docente:
          type: string
          example: "Dr. Carlos Pérez"
        horario:
          type: string
          example: "Lunes y Miércoles 08:00-10:00"
        salon:
          type: string
          example: "Bloque A - Aula 201"
        estadoInscripcion:
          type: string
          enum: [ACTIVA, CANCELADA, RETIRADA]
          example: "ACTIVA"
    CalificacionFinal:
      type: object
      description: "Final grade record for a student in a course/period"
      required:
        - codigoAsignatura
        - nombreAsignatura
        - periodo
        - creditos
        - notaFinal
        - estado
        - fechaRegistro
        - docente
      properties:
        codigoAsignatura:
          type: string
          example: "IS-201"
        nombreAsignatura:
          type: string
          example: "Programación Orientada a Objetos"
        periodo:
          type: string
          example: "2024-2"
        creditos:
          type: integer
          example: 3
        notaFinal:
          type: number
          format: float
          minimum: 0.0
          maximum: 5.0
          example: 3.8
        notaHabilitacion:
          type: number
          format: float
          minimum: 0.0
          maximum: 5.0
          nullable: true
          description: "Make-up / habilitación grade, if applicable"
          example: null
        estado:
          type: string
          enum: [APROBADO, REPROBADO, CANCELADO]
          example: "APROBADO"
        fechaRegistro:
          type: string
          format: date
          example: "2024-12-10"
        docente:
          type: string
          example: "Dr. Carlos Pérez"
...
  responses:
    Unauthorized:
      description: "Missing or invalid Bearer token"
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            detail:
              error: "UNAUTHORIZED"
              message: "Token has expired"
              timestamp: "2025-07-10T14:30:00.000Z"
              requestId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    Forbidden:
      description: "Insufficient OAuth 2.0 scope"
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            detail:
              error: "FORBIDDEN"
              message: "Insufficient scope. Required: 'academico:read'."
              timestamp: "2025-07-10T14:30:00.000Z"
              requestId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    NotFound:
      description: "Resource not found"
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                $ref: "#/components/schemas/ErrorResponse"
          example:
            detail:
              error: "NOT_FOUND"
              message: "Programme 'NOT-EXIST' not found."
              timestamp: "2025-07-10T14:30:00.000Z"
              requestId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
    TooManyRequests:
      description: "Rate limit exceeded"
      headers:
        Retry-After:
          schema:
            type: integer
          description: "Seconds until the rate limit resets"
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                $ref: "#/components/schemas/ErrorResponse"

# =============================================================================
# Parameters (reusable)
# =============================================================================
...
paths:
...
  /estudiantes/{estudianteId}/inscripciones:
    get:
      tags: [Estudiantes]
      operationId: getInscripciones
      summary: "Obtener inscripciones del estudiante"
      description: |
        Returns the enrollment list for the given student.
        Optionally filter by `periodo`.

        **Required scope:** `estudiantes:read`
      security:
        - BearerAuth: [estudiantes:read]
        - OAuth2ClientCredentials: [estudiantes:read]
        - OAuth2AuthorizationCode: [estudiantes:read]
      parameters:
        - name: estudianteId
          in: path
          required: true
          description: "Student identifier (e.g. EST-TEST-001)"
          schema:
            type: string
            example: "EST-TEST-001"
        - name: periodo
          in: query
          required: false
          description: "Filter by period code (e.g. 2025-2)"
          schema:
            type: string
            example: "2025-2"
      responses:
        "200":
          description: "List of enrollments"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inscripcion"
              example:
                - codigoAsignatura: "IS-301"
                  nombreAsignatura: "Ingeniería de Software I"
                  codigoPrograma: "ING-SIS-01"
                  periodo: "2025-2"
                  creditos: 3
                  grupo: "A"
                  docente: "Dr. Carlos Pérez"
                  horario: "Lunes y Miércoles 08:00-10:00"
                  salon: "Bloque A - Aula 201"
                  estadoInscripcion: "ACTIVA"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /estudiantes/{estudianteId}/calificaciones/{periodo}:
    get:
      tags: [Estudiantes]
      operationId: getCalificaciones
      summary: "Obtener calificaciones por período"
      description: |
        Returns the final grades for the given student in the specified academic period.

        **Required scope:** `estudiantes:read`
      security:
        - BearerAuth: [estudiantes:read]
        - OAuth2ClientCredentials: [estudiantes:read]
        - OAuth2AuthorizationCode: [estudiantes:read]
      parameters:
        - name: estudianteId
          in: path
          required: true
          description: "Student identifier"
          schema:
            type: string
            example: "EST-TEST-001"
        - name: periodo
          in: path
          required: true
          description: "Academic period code (YYYY-N)"
          schema:
            type: string
            example: "2024-1"
      responses:
        "200":
          description: "List of final grades"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalificacionFinal"
              example:
                - codigoAsignatura: "IS-101"
                  nombreAsignatura: "Fundamentos de Programación"
                  periodo: "2024-1"
                  creditos: 3
                  notaFinal: 4.2
                  notaHabilitacion: null
                  estado: "APROBADO"
                  fechaRegistro: "2024-06-10"
                  docente: "Dra. María López"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"